Modify ItemManager to handle "Buy" command Can now Buy item and add to inventory and show balance

+++ Item.cpp
@@ -4,7 +4,7 @@ Item::Item()
 {
 }
 
-Item::Item(std::string _name, double _price)
+Item::Item(std::string _name, int _price)
 {
     name = _name;
     price = _price;
@@ -15,7 +15,7 @@ std::string Item::getName()
     return name;
 }
 
-double Item::getPrice()
+int Item::getPrice()
 {
     return price;
 }
